Incorporate requested changes
