+++ package.json
@@ -65,9 +65,10 @@
                 "icon": "$(add)"
             },
             {
-                "command": "containerApps.deployImage",
-                "title": "%containerApps.deployImage%",
-                "category": "Azure Container Apps"
+                "command": "containerApps.deploy",
+                "title": "%containerApps.deploy%",
+                "category": "Azure Container Apps",
+                "icon": "$(cloud-upload)"
             },
             {
                 "command": "containerApps.deployImageApi",
@@ -148,12 +149,6 @@
                 "command": "containerApps.addScaleRule",
                 "title": "%containerApps.addScaleRule%",
                 "category": "Azure Container Apps"
-            },
-            {
-                "command": "containerApps.deploy",
-                "title": "%containerApps.deploy%",
-                "category": "Azure Container Apps",
-                "icon": "$(cloud-upload)"
             }
         ],
         "menus": {
@@ -181,7 +176,7 @@
                     "group": "3@1"
                 },
                 {
-                    "command": "containerApps.deployImage",
+                    "command": "containerApps.deploy",
                     "when": "view == azureResourceGroups && viewItem =~ /containerApp[^s]/i",
                     "group": "3@1"
                 },
